backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "relevancy-backend"
version = "0.1.0"
description = "Relevancy FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "sentence-transformers>=3.0",
    "pymupdf4llm>=0.0.17",
    "python-pptx>=1.0",
    "python-docx>=1.1",
    "python-multipart>=0.0.9",
    "numpy>=1.26",
    "anki-parser",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
